backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
faiss-cpu==1.9.0.post1
sentence-transformers==2.2.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
